pyproject.toml
[project]
name = "task-tracker-cli"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "prompt-toolkit (>=3.0.51,<4.0.0)"
]

packages = [
    {include = "task_tracker_cli", from = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
task-cli = "task_tracker_cli.cli:cli"
